refactor: split package audit tooling into modules
